pyproject.toml
[project]
name = "matmoms-tracker"
version = "0.1.0"
description = "Track Swedish grocery prices to measure VAT pass-through after April 1 2026"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.49",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pyyaml>=6.0",
    "typer>=0.15",
    "httpx>=0.28",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
matmoms = "matmoms.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
